Invoke pgxc_pool_reload() after adding/deleting nodes via pgxc_ctl
pgxc_ctl provides nice interface to add/remove coordinators/datanodes. But the
user is left with an important step of reloading the pooler information on the
nodes. This results in various issues, including the INSERT failures as
reported on the ML. This fix ensures that pgxc_ctl automatically reloads pooler
information on the nodes after cluster topology changes.