Fix reference to "node id" where we really should be looking at nodeOid
While executing ALTER NODE we check if the node being altered is marked as
a "primary" datanode and fail if there already exists another datanode marked
as "primary". But while checking for self, we compare with primary_data_node
which is really an OID of the node and not the node id. Fix that so that we
compare with the right value