Add basic support to save pgxc_ctl command history to a file and read it back next time
This code assumes GNU History library is available. We don't make any checks
for the same though because the existing code did not do that either and it was
using some other history functions anyways. The history is saved to
$PGXC_CTL_HOME/.pgxc_ctl_history and there is no way to change that right now.
In future, we may want to add support to override the default through
environment variable etc. PostgreSQL's src/bin/psql/input.c has some examples
about how to do this