Refactor GTM connection management.
1. GTM keeps track of currently open transactions in a global data structure.
When a specific client disconnects from GTM, it needs to figure out if any open
transactions should be reclaimed or not. Since there is one GTM thread per
client connection, we were using pthread_id() to record this information such
that when a connection ends, all transactions with gti_thread_id ==
pthread_self() are removed. This works OK in a standalone GTM. But when things
are failed over to a standby, the pthread_id on the old master makes no sense
on the new master. So now we assign a monotonically increasing identifier to
each incoming connection. Proxy remembers this identifier and sends it back to
the new master when it reconnects. This identifier is also backed up to the
standby. Identifier wrapped around is handled, but given an identifier is
assigned for every new connection, wrap around is almost near impossible to
happen in real world.

2. The conn_id which proxy uses to identify its client connection quickly while
handling responses from GTM, is an index into an array of client connections
that proxy maintains on a per thread basis. Currently, when a client
disconnects from the proxy, we were moving connections around in that array so
as to keep it packed. But this can mess up things if there are outstanding
requests on the GTM since it will encode the old conn_id in the message
response which has now changed on the proxy. We fix this by not moving
connections around. So an array index once assigned remains the same. But to
quickly iterate through the list of open connections, we also maintain a
condensed array which acts like a map to the actual connections.

3. There was a long standing bug report about GTM running out of threads. This
was kind of surprising because even though GTM forks one thread per connection,
when clients are connected via proxy, the number of GTM threads should be
pretty small. We noticed that a call to pthread_detach() could have been called
on a thread id which is not yet set and hence thread specific resources might
not get reclaimed when a thread exits. We don't have a reproducible scenario
for this bug, but this looks like a fairly decent theory to explain the bug. So
this patch fixes that issue.

4. GTM and GTM proxy code assumes that the first few members of GTM_ThreadInfo
and GTMProxy_ThreadInfo structure have the same storage layout. This was very
easy to get messed up, like we saw during this patch development. So those
common members are now moved to #define and included via that mechanism in both
the structures. We could have instead also defined a common structure, but
unfortunately that could lead to lot of code changes. So this seems like a good
compromise for now.