Add support for materialized views.
Right now we only support MatViews on coordinators. They are not even created
on the datanodes, though when refreshed they will be refreshed on all the
coordinators.

Also extend LOCAL TEMP tables to mean temp tables created only on the
coordinator. This allows us to use the MatView's mechanism of refresh, though
in the longer run we may want to revisit this and see if the refresh mechanism
needs to be dealt differently in XL, especially if we end up supporting
MatViews on the datanodes.

Finally, slightly modify the join push-down so that a table without
distribution mechanism are assumed coordinator only table and joined locally.
This has a side-effect that pg_catalog tables are now also joined on the
coordinator. This needs to be fixed.