Do not propogate REINDEX command to coordinators for indexes on temporary tables
We don't create temp tables on the remote coordinators and hence we need not
(in fact, must not) send REINDEX command to those coordinators. Similarly, for
indexes on materialized views, REINDEX command should only be sent to the
coordinators. We now look at the underlying table properties and decide the
target node based on that. This is in sync with the CREATE TABLE/VIEW treatment
for these objects.