Do not run queries starting from a remote coordinator inside a transaction block
There are only a few occasions where a remote coordinator may need to connect
to datanodes, CREATE MATERIALIZED VIEW is one such example. We don't the remote
coordinator to run the SELECT query inside a transaction block. The DDL on the
remote coordinator is still done inside a transaction block and the SELECT uses
the global snapshot so the results will be consistent across all coordinators.
But using auto commit helps us running 2PC from the remote coordinator which
may confuse the datanodes