Turn tables in updatable_views testcase to replicated tables
This test contains many INSERT/UPDATE/DELETE on the first column, also a
primary key, of the table. Since XL by default distributes a table on its
primary key and it does not support INSERT/UPDATE/DELETE on the distribution
column, the test case fails with many errors. So we instead use replicated
tables in the test case. There are still a few failures that need to be
investigated.