Do not error out if transaction type is INTERNALOID for aggregates
Its not clear why this was added in the original commit. Neither the commit
message nor the code has any comment explaining this. Moreover as clear from
the regression tests, we do support transaction types as INTERNALOID. Sure
aggregates can't support 2-step aggregation for internal types because the
transition function result needs to be sent to the coordinator for collection.
But that's a separate patch, if at all needed