User GXID instead of Transaction Handle while sending backup actions to the GTM standby
Transaction handles are indexes into a global array of GTM_TransactionInfo
structures. So they are very local to the GTM master. While there was an idea
to send them back to the client as identifiers for quick look-up in further
communications, we never traded that path. So it does not makes sense to back
them up on the standbys.