Force XID assignment before a snapshot is obtained on a local coordinator
Without an XID, GTM does not have a way to track xmin of every open snapshot
and that can lead to early advancing of RecentGlobalXmin.

We need a mechanism to track such snapshots over their lifetime. But till that
infrastructure is added, we force XID assignment at the local coordinator. We
are still being a bit lax about doing it on a datanode when a snapshot is
directly requested by a datanode. Keeping watch for any side-effects of that
until a proper infrastructure is added