Fix issues with ANALYZE command
ANALYZE command was failing to fetch the relation statistics such as relpages
and reltuples from the datanodes. This could result gross planner errors and
generation of very bad plans, as seen in DBT3 benchmark. We now pull relation
stats from the datanodes at ANALYZE time, just the way we do at VACUUM time.
There was also a lingering bug where the results of ANALYZE command on a
datanode won't be available to the coordinator. This turned out to be because
of using a stale snapshot and hence coordinator fails to see the new data
generated by previous ANALYZE command. We fix that by taking a fresh snapshot
at the coordinator and sending that down to the datanodes