Use a linked list to track the session and local SET parameters
Earlier we were using a hash table to track these parameters. While thats great
to quick lookup and insert, as regression tests show, its important to send them
in the same order to the remote nodes. Thats not possible using a hash table.
So use a list instead. We don't expect a large number of parameters to be
active at any point and hence list should not perform too badly either