Do not use READ ONLY transaction while dumping data using pg_dump
We use nextval(sequence) to get a consistent sequence value directly from the
GTM, since sequence values could be catched at different coordinators. But that
requires a RW transaction. Its not ideal for pg_dump to use RW transaction, but
its not terrible either given that its run in a very controlled manner. So
change it that way until we find a more elegant solution

Also fix some assorted issues with pg_dump. It now seems to pass on the
"regression" database after a round of regression run