Do not silently skip FK constraints if loose_constraints are ON.
While the git history and comments do not tell us much, it seems this was done
to avoid creating FK constraints which may not be correctly enforced when
loose_constraints is ON (which allows user to define globally non-enforcable
constraints). But instead of wholesale skipping of FKs, it seems we should
rather let the users determine which constraints can be enforced and allow them
to define them.